requirements.txt

mysqlclient==1.3.13
beautifulsoup4==4.6.3
django-heroku==0.3.1